update to use renamed to deprecated
